[Scenario] Ensure Timeline Inspector stays in sync after Save All
